requirements.txt
matplotlib==3.10.8
nibabel==5.3.3
numpy==2.4.0
pandas==2.3.3
py_cpuinfo==9.0.0
PyYAML==6.0.3
radiomics==0.1
scikit_learn==1.8.0
seaborn==0.13.2
sympy==1.13.3
torch==2.7.1+cu118
